validator and contraint

+++ src/routes/deletePokemon.js
@@ -4,14 +4,22 @@ module.exports = (app) => {
     app.delete('/api/pokemons/:id', (req, res) => {
         Pokemon.findByPk(req.params.id)
             .then(pokemon => {
-                const pokemonDeleted = pokemon;
-                Pokemon.destroy({
+                if (pokemon === null) {
+                    const message = `Le pokémon demandé n'existe pas. Réessayez avec un autre identifiant.`
+                    return res.status(404).json({ message });   
+                }
+                
+                return Pokemon.destroy({
                     where: { id: pokemon.id}
                 })
                 .then(_ => {
-                    const message = `Le pokémon avec l'identifiant n°${pokemonDeleted.id} a bien été supprimé.`;
-                    res.json({ message, data: pokemonDeleted });
+                    const message = `Le pokémon avec l'identifiant n°${pokemon.id} a bien été supprimé.`;
+                    res.json({ message, data: pokemon });
                 })
+            })
+            .catch(error => {
+                const message = `Le pokémon n'a pas pu être supprimé. Réessayez dans quelques instants.`
+                res.status(500).json({ message, data: error }); 
             })
     })
 }
